refactor: inline E.164 string rule
